:arrow_up: omnisharp-client 7.2.2
This change updates omnisharp-client to 7.2.2 to take a new fix to
their server script which causes a bundled Mono 5.2 to be used on
macOS and Linux.  This resolves issue #7 where OmniSharp does not
start up on Linux when an older Mono version is installed.  The
minimum supported Mono version for OmniSharp is 5.2.

This also includes a fix where their language server wasn't picking
up the --zero-based-indices command line flag which caused most
language server operations to give incorrect results.